requirements.txt
#check week-9-complete code to add versions and more packages
pandas
openpyxl
flask
flask-sqlalchemy
Flask-WTF
flask-wtf
flask-marshmallow
marshmallow-sqlalchemy
requests
selenium
numpy
pytest==7.2.1
pytest-cov==4.0.0
pytest-flask==1.2.0


#flask-login